Create a header utility routine
Create a routine that can help process a list separated by a given
character when the elements of that list may have that separator
within double-quoted sections.  This will be used both to parse the
multiple options of an Accept header, as well as to parse the
parameters of an individual content type.