Add a version aliasing mechanism
Adds a way to set up aliases for versions.  This is intended for
ensuring canonical values in the face of type substitutions.
Consider for instance a type rule that yields version "v1.1",
and we want to map this request to the v2 api.  By adding
"alias.v1.1 = v2" to the configuration, the version number
used (and present in the "aversion.version" WSGI environment
variable) will be "v2" rather than "v1.1".